Refactor family crest
