+++ src/app/family-crest/family-crest-store.ts
@@ -9,10 +9,3 @@ export const useFamilyCrestStore = create<FamilyCrestState>((set) => ({
   crest: null,
   updateCrest: (crest) => set({ crest }),
 }));
-
-export const useFamilyCrest = () => {
-  return {
-    crest: useFamilyCrestStore((s) => s.crest),
-    updateCrest: useFamilyCrestStore((s) => s.updateCrest),
-  };
-};
