+++ src/app/family-crest/useGenerateCrest.ts
@@ -1,9 +1,9 @@
 import { useMutation } from "@tanstack/react-query";
 import { FormSchema as familyCrestFormSchema } from "./family-crest-form";
-import { useFamilyCrest } from "./family-crest-store";
+import { useFamilyCrestStore } from "./family-crest-store";
 
 export default function useGenerateCrest() {
-  const { updateCrest } = useFamilyCrest();
+  const updateCrest = useFamilyCrestStore((s) => s.updateCrest);
   const { mutate, isPending, error } = useMutation({
     mutationFn: async (family: familyCrestFormSchema) => {
       const response = await fetch("/api/generate-family-crest", {
